+++ JSQMessagesViewController/Views/JSQMessagesTypingView.h
@@ -1,0 +1,28 @@
+//
+//  Created by Jesse Squires
+//  http://www.jessesquires.com
+//
+//
+//  Documentation
+//  http://cocoadocs.org/docsets/JSQMessagesViewController
+//
+//
+//  GitHub
+//  https://github.com/jessesquires/JSQMessagesViewController
+//
+//
+//  License
+//  Copyright (c) 2014 Jesse Squires
+//  Released under an MIT license: http://opensource.org/licenses/MIT
+//
+
+#import <UIKit/UIKit.h>
+
+@interface JSQMessagesTypingView : UIView
+
+@property (strong, nonatomic) UIColor *dotsColor;
+@property (strong, nonatomic) UIColor *animateToColor;
+@property (assign, nonatomic) CGFloat animationDuration;
+@property (assign, nonatomic) BOOL animated;
+
+@end